+++ tests/Plugin/Wechat/Extend/Complaints/QueryDetailPluginTest.php
@@ -48,4 +48,29 @@ class QueryDetailPluginTest extends TestCase
             '_service_url' => 'v3/merchant-service/complaints-v2/yansongda',
         ], $result->getPayload()->all());
     }
+
+    public function testNormalWithEncryptedContents()
+    {
+        $payload = [
+            "complaint_id" => "yansongda",
+        ];
+
+        $rocket = new Rocket();
+        $rocket->setPayload(new Collection($payload));
+
+        $result = $this->plugin->assembly($rocket, function ($rocket) {
+            $rocket->setDestination(new Collection([
+                'payer_phone' => 'WIesmK+dSJycwdhTTkNmv0Lk2wb9o7NGODovccjhyotNnRkEeh+sxRK1gNSRNMJJgkQ30m4HwcuweSO24mehFeXVNTVAKFVef/3FlHnYDZfE1c3mCLToEef7e8J/Z8TwFH1ecn3t+Jk9ZaBpQKNHdQ0Q8jcL7AnL48h0D9BcZxDekPqX6hNnKfISoKSv4TXFcgvBLFeAe4Q3KM0Snq0N5IvI86D9xZqVg6mY+Gfz0782ymQFxflau6Qxx3mJ+0etHMocNuCdgctVH390XYYMc0u+V2FCJ5cU5h/M/AxzP9ayrEO4l0ftaxL6lP0HjifNrkPcAAb+q9I67UepKO9iGw==',
+            ]));
+
+            return $rocket;
+        });
+
+        self::assertEquals([
+            '_method' => 'GET',
+            '_url' => 'v3/merchant-service/complaints-v2/yansongda',
+            '_service_url' => 'v3/merchant-service/complaints-v2/yansongda',
+        ], $result->getPayload()->all());
+        self::assertEquals('yansongda', $result->getDestination()->all()['payer_phone']);
+    }
 }